make SnapshotFileGroup an static inner class of ChangeFilesUtil
